pyproject.toml
[project]
name = "cortex"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13,<3.14"
dependencies = [
    "duckdb>=1.4.4",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langgraph>=0.2",
    "pandas>=1.4.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.0",
    "streamlit>=1.40",
]
